Commit5
Circuit to unlock the door when the correct password is entered by the user, otherwise blow an alarm to warn. Always the LCD will show the entered password with the '*' symbol. When the correct password is entered, the message "Valid Password" is displayed. Door Unlocked", and drives a DC motor, displays the message "Door Locked" after 3 seconds. When a wrong password is entered, the message "Invalid Password" appears. Locked!!", and a buzzer as a warning. executes